Resolved pedantic warnings in sort()

+++ headers/functions.h
@@ -46,9 +46,9 @@ void set_env_variable(std::string v, std::string value);
 template <typename T>
 void sort(std::vector<T>* v)
 {
-	for(int i=0; i<v->size()-1; i++)
+	for(long unsigned int i=0; i<v->size()-1; i++)
 	{
-		for(int j=i+1; j<v->size(); j++)
+		for(long unsigned int j=i+1; j<v->size(); j++)
 		{
 			if((*v)[j] < (*v)[i])
 			{
